+++ packages/chain-adapters/src/utils/index.ts
@@ -5,9 +5,6 @@ import {
   ChainId,
   fromChainId
 } from '@shapeshiftoss/caip'
-import { Status, TransferType } from '@shapeshiftoss/unchained-client'
-
-import { TxStatus, TxType } from '../types'
 
 export * from './bip44'
 export * from './utxoUtils'
@@ -18,31 +15,18 @@ export const getAssetNamespace = (type: string): AssetNamespace => {
   throw new Error(`Unknown asset namespace. type: ${type}`)
 }
 
-export const getStatus = (status: Status): TxStatus => {
-  if (status === Status.Pending) return TxStatus.Pending
-  if (status === Status.Confirmed) return TxStatus.Confirmed
-  if (status === Status.Failed) return TxStatus.Failed
-
-  return TxStatus.Unknown
-}
-
-export const getType = (type: TransferType): TxType => {
-  if (type === TransferType.Send) return TxType.Send
-  if (type === TransferType.Receive) return TxType.Receive
-
-  return TxType.Unknown
-}
-
 export const chainIdToChainLabel = (chainId: ChainId): string => {
   const { chainNamespace, chainReference } = fromChainId(chainId)
 
   switch (chainNamespace) {
     case CHAIN_NAMESPACE.Bitcoin:
       switch (chainReference) {
-        case CHAIN_REFERENCE.DogecoinMainnet:
-          return 'dogecoin'
         case CHAIN_REFERENCE.BitcoinMainnet:
           return 'bitcoin'
+        case CHAIN_REFERENCE.DogecoinMainnet:
+          return 'dogecoin'
+        case CHAIN_REFERENCE.LitecoinMainnet:
+          return 'litecoin'
         default:
           throw new Error(
             `chainReference: ${chainReference}, not supported for chainNamespace: ${chainNamespace}`