+++ packages/swapper/src/swappers/cow/cowBuildTrade/cowBuildTrade.test.ts
@@ -1,6 +1,7 @@
+import { Asset } from '@shapeshiftoss/asset-service'
 import { ethereum, FeeDataEstimate } from '@shapeshiftoss/chain-adapters'
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
-import { Asset, KnownChainIds } from '@shapeshiftoss/types'
+import { KnownChainIds } from '@shapeshiftoss/types'
 import Web3 from 'web3'
 
 import { BuildTradeInput } from '../../../api'
@@ -10,7 +11,7 @@ import { ETH, FOX, WBTC, WETH } from '../../utils/test-data/assets'
 import { CowSwapperDeps } from '../CowSwapper'
 import { CowTrade } from '../types'
 import { cowService } from '../utils/cowService'
-import { CowSwapQuoteApiInput } from '../utils/helpers/helpers'
+import { CowSwapSellQuoteApiInput } from '../utils/helpers/helpers'
 import { cowBuildTrade } from './cowBuildTrade'
 
 jest.mock('@shapeshiftoss/chain-adapters')
@@ -72,7 +73,7 @@ const feeData: FeeDataEstimate<KnownChainIds.EthereumMainnet> = {
   }
 }
 
-const expectedApiInputWethToFox: CowSwapQuoteApiInput = {
+const expectedApiInputWethToFox: CowSwapSellQuoteApiInput = {
   appData: '0x0000000000000000000000000000000000000000000000000000000000000000',
   buyToken: '0xc770eefad204b5180df6a14ee197d99d808ee52d',
   from: 'address11',
@@ -84,7 +85,7 @@ const expectedApiInputWethToFox: CowSwapQuoteApiInput = {
   validTo: 1656797787
 }
 
-const expectedApiInputWbtcToWeth: CowSwapQuoteApiInput = {
+const expectedApiInputWbtcToWeth: CowSwapSellQuoteApiInput = {
   appData: '0x0000000000000000000000000000000000000000000000000000000000000000',
   buyToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
   from: 'address11',
@@ -154,7 +155,6 @@ describe('cowBuildTrade', () => {
       sellAmount: '11111',
       sendMax: true,
       sellAssetAccountNumber: 1,
-      buyAssetAccountNumber: 2,
       wallet: <HDWallet>{},
       receiveAddress: ''
     }
@@ -181,7 +181,6 @@ describe('cowBuildTrade', () => {
       sellAmount: '1000000000000000000',
       sendMax: true,
       sellAssetAccountNumber: 0,
-      buyAssetAccountNumber: 0,
       wallet: <HDWallet>{},
       receiveAddress: ''
     }
@@ -228,7 +227,6 @@ describe('cowBuildTrade', () => {
       sellAmount: '100000000',
       sendMax: true,
       sellAssetAccountNumber: 0,
-      buyAssetAccountNumber: 0,
       wallet: <HDWallet>{},
       receiveAddress: ''
     }