+++ packages/unchained-client/src/evm/ethereum/parser/index.ts
@@ -1,7 +1,8 @@
-import { ASSET_REFERENCE, toAssetId } from '@shapeshiftoss/caip'
+import { ethAssetId } from '@shapeshiftoss/caip'
 
 import { Tx } from '../../../generated/ethereum'
 import { BaseTransactionParser, TransactionParserArgs } from '../../parser'
+import * as cowswap from './cowswap'
 import * as foxy from './foxy'
 import * as thor from './thor'
 import * as uniV2 from './uniV2'
@@ -13,11 +14,7 @@ export class TransactionParser extends BaseTransactionParser<Tx> {
   constructor(args: TransactionParserArgs) {
     super(args)
 
-    this.assetId = toAssetId({
-      chainId: this.chainId,
-      assetNamespace: 'slip44',
-      assetReference: ASSET_REFERENCE.Ethereum
-    })
+    this.assetId = ethAssetId
 
     // due to the current parser logic, order here matters (register most generic first to most specific last)
     // weth and yearn have the same sigHash for deposit(), but the weth parser is stricter resulting in faster processing times
@@ -27,7 +24,8 @@ export class TransactionParser extends BaseTransactionParser<Tx> {
       new weth.Parser({ chainId: this.chainId, provider: this.provider }),
       new uniV2.Parser({ chainId: this.chainId, provider: this.provider }),
       new thor.Parser({ chainId: this.chainId, rpcUrl: args.rpcUrl }),
-      new zrx.Parser()
+      new zrx.Parser(),
+      new cowswap.Parser()
     ])
   }
 }