+++ packages/chain-adapters/CHANGELOG.md
@@ -1,3 +1,26 @@
+# [@shapeshiftoss/chain-adapters-v7.13.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v7.13.0...@shapeshiftoss/chain-adapters-v7.13.1) (2022-07-29)
+
+# [@shapeshiftoss/chain-adapters-v7.13.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v7.12.0...@shapeshiftoss/chain-adapters-v7.13.0) (2022-07-29)
+
+
+### Features
+
+* litecoin ([#914](https://github.com/shapeshift/lib/issues/914)) ([5af5c60](https://github.com/shapeshift/lib/commit/5af5c60c954ae0295d4cd175b98da3732d99d9e4))
+
+# [@shapeshiftoss/chain-adapters-v7.12.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v7.11.2...@shapeshiftoss/chain-adapters-v7.12.0) (2022-07-26)
+
+
+### Features
+
+* cowswap tx parser ([#909](https://github.com/shapeshift/lib/issues/909)) ([4e67743](https://github.com/shapeshift/lib/commit/4e67743339b916e87903c44a2d4eccf23b4b3a59))
+
+# [@shapeshiftoss/chain-adapters-v7.11.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v7.11.1...@shapeshiftoss/chain-adapters-v7.11.2) (2022-07-26)
+
+
+### Bug Fixes
+
+* unrug lib bump packages ([#910](https://github.com/shapeshift/lib/issues/910)) ([c914e58](https://github.com/shapeshift/lib/commit/c914e58a2832e5bc196d062074499ec46a200d50))
+
 # [@shapeshiftoss/chain-adapters-v7.11.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v7.11.0...@shapeshiftoss/chain-adapters-v7.11.1) (2022-07-25)
 
 # [@shapeshiftoss/chain-adapters-v7.11.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v7.10.0...@shapeshiftoss/chain-adapters-v7.11.0) (2022-07-25)